README: Use lower-case references in binary calls
I've found that on some file systems (at least in Travis), case-sensitivity causes problems in not finding `FRS-replace` as a binary (it gets stored as `frs-replace`, as per the npm package name and only works with that). In this PR, I also changed it in `require` statements (not sure whether this is needed on those systems, but makes sense I think to reflect the binary and npm package name) as well though not when just referencing the project name.